[ :recycle: ] | refact: module - matrix > math  (module::matrix::math)
